Move allocator into nested workspace
This allows rustgc to import just the allocator and use it in the
standard library. When programs which use libgc are compiled with
rustgc, the nested allocator crate is not compiled, instead an
allocator which was pre-compiled with rustgc-libstd will be used.

For this to work, programs which were not compiled with rustgc must use
libgc's "standalone" feature flag. This isn't intended to be used
regularly, however, it is a useful feature for development because it
lets us test libgc features without recompiling rustgc.
