+++ .buildbot.sh
@@ -12,18 +12,8 @@ sh rustup.sh --default-host x86_64-unknown-linux-gnu \
     --profile minimal \
     -y
 export PATH=`pwd`/.cargo/bin/:$PATH
-cargo check
+cargo check --features "standalone"
 
 rustup toolchain install nightly --allow-downgrade --component rustfmt
 cargo +nightly fmt --all -- --check
 
-# Build and test with rustgc
-git clone https://github.com/softdevteam/rustgc
-mkdir -p rustgc/build/rustgc
-(cd rustgc && ./x.py build --config ../.buildbot.config.toml)
-
-rustup toolchain link rustgc rustgc/build/x86_64-unknown-linux-gnu/stage1
-
-cargo clean
-
-cargo +rustgc test --features "rustgc" -- --test-threads=1
