+++ Cargo.toml
@@ -7,7 +7,7 @@ edition = "2018"
 [features]
 # Enable this feature to turn on additional GC optimizations that are only
 # possible with the rustgc fork of the compiler.
-rustgc = []
+standalone = ["allocator"]
 
 # Enable various GC based statistics. Stats are disabled by default as they have
 # a run-time cost and are expected to only be used for profiling purposes.
@@ -15,8 +15,12 @@ gc_stats = []
 
 [dependencies]
 libc = "*"
+allocator = { path = "allocator", optional = true }
 
 [build-dependencies]
 rerun_except = "0.1"
 num_cpus = "1.12"
 which = "4.0"
+
+[workspace]
+members = ["allocator"]
