+++ allocator/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "allocator"
+version = "0.1.0"
+authors = ["Jake Hughes <jh@jakehughes.uk>"]
+edition = "2018"
+license = "Apache-2.0 OR MIT"
+
+[features]
+# Enable this feature to turn on additional GC optimizations that are only
+# possible with the rustgc fork of the compiler.
+rustgc = ["core", "compiler_builtins"]
+
+[dependencies]
+core = { version = "1.0.0", optional = true, package = 'rustc-std-workspace-core' }
+compiler_builtins = { version = "0.1.10", optional = true, features = ['rustc-dep-of-std'] }
